Suggested CCSS text. (#52)
* Suggested CCSS text.

* Update data-access-unit.md

fixed a small typo

* Update data-access-unit.md

A) clarified the second sentence of point 3 to indicate that not all bacterial sequences have been mapped. Instead just say that there are 'instances' of bacterial sequence mapping.
B) updated point 4 to indicated that we will be also releasing the updated 'BAM' files (along with the gCVF files).