Mtreadwa/data browser updates data transfer app (#54)
* Data Browser & Data Transfer App Updates

Added content to the Data Transfer App page. Waiting on updates to the app's color scheme to add new pngs and gifs.

Outlined updated Making a Data Request page for the new Data Browser. Waiting on Snaggit and completion of minor browser edits to make gifs.

* Update data-request.md

Continued working on the text.

* gifs and pngs added

Added gifs and images to How to Make a Data Request page

* Edits to gifs/pngs

Incorporated Kirby's feedback related to images and gifs

* Updated gif and png in Data Transfer App page

Added gifs and pngs with the new color scheme to the Data Transfer App page.

* Fixed some typos