Genome Paint one pager
As per Alex's recomendation, the one pager contains an overview of the tool, description of data displayed, and brief description of the 3 views (cohort, sample, matrix). Content is roughly from the long genome paint tutorial and Alex's presentation. I added an overview image of the landing page to match the format of what we have for protein paint. Also added some gifs to show how to navigate and move among different views.