Remove studies mention
Studies are going to be part of visualization community and no longer need to be mentioned as part of the portal. A section in the docs for Visualization Community does not make sense at the moment since there is no ability to create your own visualizations. This can be revisted in v2.